+++ bin/faucet/src/client.rs
@@ -19,7 +19,7 @@ use miden_objects::{
     },
     note::{Note, NoteType},
     transaction::{
-        ChainMmr, ExecutedTransaction, ForeignAccountInputs, InputNote, InputNotes,
+        ExecutedTransaction, ForeignAccountInputs, InputNote, InputNotes, PartialBlockChain,
         TransactionArgs, TransactionScript,
     },
     utils::Deserializable,
@@ -61,7 +61,7 @@ impl FaucetClient {
     /// If the faucet account is not found on chain, it will be created on submission of the first
     /// minting transaction.
     pub async fn new(config: &FaucetConfig) -> Result<Self, ClientError> {
-        let (mut rpc_api, root_block_header, root_chain_mmr) =
+        let (mut rpc_api, root_block_header, root_partial_block_chain) =
             initialize_faucet_client(config).await?;
 
         let faucet_account_data = AccountFile::read(&config.faucet_account_path)
@@ -98,7 +98,7 @@ impl FaucetClient {
             faucet_account,
             faucet_account_data.account_seed,
             root_block_header,
-            root_chain_mmr,
+            root_partial_block_chain,
         ));
 
         let public_key = match &faucet_account_data.auth_secret_key {
@@ -209,7 +209,7 @@ impl FaucetClient {
 /// Initializes the faucet client by connecting to the node and fetching the root block header.
 pub async fn initialize_faucet_client(
     config: &FaucetConfig,
-) -> Result<(ApiClient<Channel>, BlockHeader, ChainMmr), ClientError> {
+) -> Result<(ApiClient<Channel>, BlockHeader, PartialBlockChain), ClientError> {
     let endpoint = tonic::transport::Endpoint::try_from(config.node_url.to_string())
         .context("Failed to parse node URL from configuration file")?
         .timeout(Duration::from_millis(config.timeout_ms));
@@ -232,15 +232,15 @@ pub async fn initialize_faucet_client(
 
     let root_block_header = root_block_header.try_into().context("Failed to parse block header")?;
 
-    let root_chain_mmr = ChainMmr::new(
+    let root_partial_block_chain = PartialBlockChain::new(
         PartialMmr::from_peaks(
             MmrPeaks::new(0, Vec::new()).expect("Empty MmrPeak should be valid"),
         ),
         Vec::new(),
     )
-    .expect("Empty ChainMmr should be valid");
+    .expect("Empty PartialBlockChain should be valid");
 
-    Ok((rpc_api, root_block_header, root_chain_mmr))
+    Ok((rpc_api, root_block_header, root_partial_block_chain))
 }
 
 /// Requests account state from the node.