+++ crates/store/src/state.rs
@@ -22,15 +22,15 @@ use miden_objects::{
     AccountError,
     account::{AccountDelta, AccountHeader, AccountId, StorageSlot},
     block::{
-        AccountTree, AccountWitness, BlockHeader, BlockInputs, BlockNumber, NullifierWitness,
-        ProvenBlock,
+        AccountTree, AccountWitness, BlockChain, BlockHeader, BlockInputs, BlockNumber,
+        NullifierTree, NullifierWitness, ProvenBlock,
     },
     crypto::{
         hash::rpo::RpoDigest,
-        merkle::{Mmr, MmrDelta, MmrError, MmrPeaks, MmrProof, PartialMmr, SmtProof},
+        merkle::{Mmr, MmrDelta, MmrProof, PartialMmr, SmtProof},
     },
     note::{NoteId, Nullifier},
-    transaction::{ChainMmr, OutputNote},
+    transaction::{OutputNote, PartialBlockChain},
     utils::Serializable,
 };
 use tokio::{
@@ -48,7 +48,6 @@ use crate::{
         GetBlockInputsError, InvalidBlockError, NoteSyncError, StateInitializationError,
         StateSyncError,
     },
-    nullifier_tree::NullifierTree,
 };
 // STRUCTURES
 // ================================================================================================
@@ -60,106 +59,19 @@ pub struct TransactionInputs {
     pub found_unauthenticated_notes: BTreeSet<NoteId>,
 }
 
-/// A [Merkle Mountain Range](Mmr) defining a chain of blocks.
-#[derive(Debug, Clone)]
-pub struct Blockchain(Mmr);
-
-impl Blockchain {
-    /// Returns a new Blockchain.
-    pub fn new(chain_mmr: Mmr) -> Self {
-        Self(chain_mmr)
-    }
-
-    /// Returns the tip of the chain, i.e. the number of the latest block in the chain.
-    pub fn chain_tip(&self) -> BlockNumber {
-        let block_number: u32 = (self.0.forest() - 1)
-            .try_into()
-            .expect("chain_mmr always has, at least, the genesis block");
-
-        block_number.into()
-    }
-
-    /// Returns the current peaks of the MMR.
-    pub fn peaks(&self) -> MmrPeaks {
-        self.0.peaks()
-    }
-
-    /// Returns the peaks of the MMR at the state specified by `forest`.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the specified `forest` value is not valid for this MMR.
-    pub fn peaks_at(&self, forest: usize) -> Result<MmrPeaks, MmrError> {
-        self.0.peaks_at(forest)
-    }
-
-    /// Adds a block commitment to the MMR. The caller must ensure that this commitent is the one
-    /// for the next block in the chain.
-    pub fn push(&mut self, block_commitment: RpoDigest) {
-        self.0.add(block_commitment);
-    }
-
-    /// Returns an [`MmrProof`] for the leaf at the specified position.
-    pub fn open(&self, pos: usize) -> Result<MmrProof, MmrError> {
-        self.0.open_at(pos, self.0.forest())
-    }
-
-    /// Returns a reference to the underlying [`Mmr`].
-    pub fn as_mmr(&self) -> &Mmr {
-        &self.0
-    }
-
-    /// Creates a [`PartialMmr`] at the state of the latest block (i.e. the block's chain root will
-    /// match the hashed peaks of the returned partial MMR). This MMR will include authentication
-    /// paths for all blocks in the provided set.
-    pub fn partial_mmr_from_blocks(
-        &self,
-        blocks: &BTreeSet<BlockNumber>,
-        latest_block_number: BlockNumber,
-    ) -> PartialMmr {
-        // Using latest block as the target forest means we take the state of the MMR one before
-        // the latest block. This is because the latest block will be used as the reference
-        // block of the batch and will be added to the MMR by the batch kernel.
-        let target_forest = latest_block_number.as_usize();
-        let peaks = self
-            .peaks_at(target_forest)
-            .expect("target_forest should be smaller than forest of the chain mmr");
-        // Grab the block merkle paths from the inner state.
-        let mut partial_mmr = PartialMmr::from_peaks(peaks);
-
-        for block_num in blocks.iter().map(BlockNumber::as_usize) {
-            // SAFETY: We have ensured block nums are less than chain length.
-            let leaf = self
-                .0
-                .get(block_num)
-                .expect("block num less than chain length should exist in chain mmr");
-            let path = self
-                .0
-                .open_at(block_num, target_forest)
-                .expect("block num and target forest should be valid for this mmr")
-                .merkle_path;
-            // SAFETY: We should be able to fill the partial MMR with data from the chain MMR
-            // without errors, otherwise it indicates the chain mmr is invalid.
-            partial_mmr
-                .track(block_num, leaf, &path)
-                .expect("filling partial mmr with data from mmr should succeed");
-        }
-
-        partial_mmr
-    }
-}
-
 /// Container for state that needs to be updated atomically.
 struct InnerState {
     nullifier_tree: NullifierTree,
-    blockchain: Blockchain,
+    blockchain: BlockChain,
     account_tree: AccountTree,
 }
 
 impl InnerState {
     /// Returns the latest block number.
     fn latest_block_num(&self) -> BlockNumber {
-        self.blockchain.chain_tip()
+        self.blockchain
+            .chain_tip()
+            .expect("chain should always have at least the genesis block")
     }
 }
 
@@ -195,7 +107,9 @@ impl State {
 
         let inner = RwLock::new(InnerState {
             nullifier_tree,
-            blockchain: Blockchain::new(chain_mmr),
+            // SAFETY: We assume the loaded MMR is valid and does not have more than u32::MAX
+            // entries.
+            blockchain: BlockChain::from_mmr_unchecked(chain_mmr),
             account_tree,
         });
 
@@ -305,21 +219,27 @@ impl State {
             }
 
             // compute update for nullifier tree
-            let nullifier_tree_update = inner.nullifier_tree.compute_mutations(
-                block.created_nullifiers().iter().map(|nullifier| (*nullifier, block_num)),
-            );
+            let nullifier_tree_update = inner
+                .nullifier_tree
+                .compute_mutations(
+                    block.created_nullifiers().iter().map(|nullifier| (*nullifier, block_num)),
+                )
+                .map_err(InvalidBlockError::NewBlockNullifierAlreadySpent)?;
 
-            if nullifier_tree_update.root() != header.nullifier_root() {
+            if nullifier_tree_update.as_mutation_set().root() != header.nullifier_root() {
                 return Err(InvalidBlockError::NewBlockInvalidNullifierRoot.into());
             }
 
             // compute update for account tree
-            let account_tree_update = inner.account_tree.compute_mutations(
-                block
-                    .updated_accounts()
-                    .iter()
-                    .map(|update| (update.account_id(), update.final_state_commitment())),
-            );
+            let account_tree_update = inner
+                .account_tree
+                .compute_mutations(
+                    block
+                        .updated_accounts()
+                        .iter()
+                        .map(|update| (update.account_id(), update.final_state_commitment())),
+                )
+                .map_err(InvalidBlockError::NewBlockDuplicateAccountIdPrefix)?;
 
             if account_tree_update.as_mutation_set().root() != header.account_root() {
                 return Err(InvalidBlockError::NewBlockInvalidAccountRoot.into());
@@ -450,7 +370,7 @@ impl State {
         if let Some(header) = block_header {
             let mmr_proof = if include_mmr_proof {
                 let inner = self.inner.read().await;
-                let mmr_proof = inner.blockchain.open(header.block_num().as_usize())?;
+                let mmr_proof = inner.blockchain.open(header.block_num())?;
                 Some(mmr_proof)
             } else {
                 None
@@ -479,7 +399,11 @@ impl State {
     #[instrument(target = COMPONENT, skip_all, ret(level = "debug"))]
     pub async fn check_nullifiers(&self, nullifiers: &[Nullifier]) -> Vec<SmtProof> {
         let inner = self.inner.read().await;
-        nullifiers.iter().map(|n| inner.nullifier_tree.open(n)).collect()
+        nullifiers
+            .iter()
+            .map(|n| inner.nullifier_tree.open(n))
+            .map(NullifierWitness::into_proof)
+            .collect()
     }
 
     /// Queries a list of [`NoteRecord`] from the database.
@@ -543,7 +467,7 @@ impl State {
         let (batch_reference_block, partial_mmr) = {
             let inner_state = self.inner.read().await;
 
-            let latest_block_num = inner_state.blockchain.chain_tip();
+            let latest_block_num = inner_state.latest_block_num();
 
             let highest_block_num =
                 *blocks.last().expect("we should have checked for empty block references");
@@ -559,10 +483,19 @@ impl State {
             // there is no need to prove its inclusion.
             blocks.remove(&latest_block_num);
 
-            (
-                latest_block_num,
-                inner_state.blockchain.partial_mmr_from_blocks(&blocks, latest_block_num),
-            )
+            // SAFETY:
+            // - The latest block num was retrieved from the inner blockchain from which we will
+            //   also retrieve the proofs, so it is guaranteed to exist in that chain.
+            // - We have checked that no block number in the blocks set is greater than latest block
+            //   number *and* latest block num was removed from the set. Therefore only block
+            //   numbers smaller than latest block num remain in the set. Therefore all the block
+            //   numbers are guaranteed to exist in the chain state at latest block num.
+            let partial_mmr = inner_state
+                .blockchain
+                .partial_mmr_from_blocks(&blocks, latest_block_num)
+                .expect("latest block num should exist and all blocks in set should be < than latest block");
+
+            (latest_block_num, partial_mmr)
         };
 
         // Fetch the reference block of the batch as part of this query, so we can avoid looking it
@@ -582,7 +515,7 @@ impl State {
             })
             .expect("DB should have returned the header of the batch reference block");
 
-        // The order doesn't matter for ChainMmr::new, so swap remove is fine.
+        // The order doesn't matter for PartialBlockChain::new, so swap remove is fine.
         let batch_reference_block_header = headers.swap_remove(header_index);
 
         // SAFETY: This should not error because:
@@ -590,13 +523,13 @@ impl State {
         // - so none of the block headers block numbers should exceed the chain length of the
         //   partial MMR,
         // - and we've added blocks to a BTreeSet, so there can be no duplicates.
-        let chain_mmr = ChainMmr::new(partial_mmr, headers)
+        let partial_block_chain = PartialBlockChain::new(partial_mmr, headers)
             .expect("partial mmr and block headers should be consistent");
 
         Ok(BatchInputs {
             batch_reference_block_header,
             note_proofs,
-            chain_mmr,
+            partial_block_chain,
         })
     }
 
@@ -673,7 +606,7 @@ impl State {
 
         let note_sync = self.db.get_note_sync(block_num, note_tags).await?;
 
-        let mmr_proof = inner.blockchain.open(note_sync.block_header.block_num().as_usize())?;
+        let mmr_proof = inner.blockchain.open(note_sync.block_header.block_num())?;
 
         Ok((note_sync, mmr_proof))
     }
@@ -724,7 +657,7 @@ impl State {
             })
             .expect("DB should have returned the header of the latest block header");
 
-        // The order doesn't matter for ChainMmr::new, so swap remove is fine.
+        // The order doesn't matter for PartialBlockChain::new, so swap remove is fine.
         let latest_block_header = headers.swap_remove(latest_block_header_index);
 
         // SAFETY: This should not error because:
@@ -732,12 +665,12 @@ impl State {
         // - so none of the block header's block numbers should exceed the chain length of the
         //   partial MMR,
         // - and we've added blocks to a BTreeSet, so there can be no duplicates.
-        let chain_mmr = ChainMmr::new(partial_mmr, headers)
+        let partial_block_chain = PartialBlockChain::new(partial_mmr, headers)
             .expect("partial mmr and block headers should be consistent");
 
         Ok(BlockInputs::new(
             latest_block_header,
-            chain_mmr,
+            partial_block_chain,
             account_witnesses,
             nullifier_witnesses,
             unauthenticated_note_proofs,
@@ -783,7 +716,18 @@ impl State {
 
         // Fetch the partial MMR at the state of the latest block with authentication paths for the
         // provided set of blocks.
-        let partial_mmr = inner.blockchain.partial_mmr_from_blocks(blocks, latest_block_number);
+        //
+        // SAFETY:
+        // - The latest block num was retrieved from the inner blockchain from which we will also
+        //   retrieve the proofs, so it is guaranteed to exist in that chain.
+        // - We have checked that no block number in the blocks set is greater than latest block
+        //   number *and* latest block num was removed from the set. Therefore only block numbers
+        //   smaller than latest block num remain in the set. Therefore all the block numbers are
+        //   guaranteed to exist in the chain state at latest block num.
+        let partial_mmr =
+            inner.blockchain.partial_mmr_from_blocks(blocks, latest_block_number).expect(
+                "latest block num should exist and all blocks in set should be < than latest block",
+            );
 
         // Fetch witnesses for all acounts.
         let account_witnesses = account_ids
@@ -797,10 +741,7 @@ impl State {
         let nullifier_witnesses: BTreeMap<Nullifier, NullifierWitness> = nullifiers
             .iter()
             .copied()
-            .map(|nullifier| {
-                let proof = inner.nullifier_tree.open(&nullifier);
-                (nullifier, NullifierWitness::new(proof))
-            })
+            .map(|nullifier| (nullifier, inner.nullifier_tree.open(&nullifier)))
             .collect();
 
         Ok((latest_block_number, account_witnesses, nullifier_witnesses, partial_mmr))